Features/call api

+++ lib/features/user_info/repository/user_info_repository.dart
@@ -1,14 +1,14 @@
 import 'dart:io';
 
+import 'package:flutter_englearn/common/utils/api_url.dart';
+import 'package:flutter_englearn/common/utils/utils.dart';
 import 'package:flutter_englearn/features/auth/repository/auth_repository.dart';
 import 'package:flutter_englearn/model/request/add_email_request.dart';
 import 'package:flutter_englearn/model/response/jwt_response.dart';
 import 'package:flutter_englearn/model/response/response_model.dart';
 import 'package:flutter_englearn/model/response/user_info_response.dart';
 import 'package:flutter_englearn/model/result_return.dart';
-import 'package:flutter_englearn/utils/const/api_url.dart';
-import 'package:flutter_englearn/utils/const/utils.dart';
-import 'package:flutter_englearn/utils/helper/helper.dart';
+import 'package:flutter_englearn/common/helper/helper.dart';
 import 'dart:developer';
 import 'package:http/http.dart' as http;
 import 'package:http_parser/http_parser.dart';
@@ -355,8 +355,7 @@ class UserInfoRepository {
         log('Token is expired', name: 'UserInfoRepository');
         await authRepository.removeJWT();
         return ResultReturn(httpStatusCode: 401, data: null);
-      }
-      else {
+      } else {
         log('Upload failed', name: 'UserInfoRepository');
         return ResultReturn(httpStatusCode: 400, data: null);
       }
@@ -386,6 +385,41 @@ class UserInfoRepository {
       if (response.statusCode == 200) {
         log('Update streak successfully', name: 'UserInfoRepository');
         return ResultReturn(httpStatusCode: 200, data: null);
+      } else if (response.statusCode == 401) {
+        log('Token is expired', name: 'UserInfoRepository');
+        await authRepository.removeJWT();
+        return ResultReturn(httpStatusCode: 401, data: null);
+      } else {
+        log('Update streak failed', name: 'UserInfoRepository');
+        return ResultReturn(httpStatusCode: 400, data: null);
+      }
+    }
+  }
+
+  Future<ResultReturn> getAvatar() async {
+    JwtResponse? jwtResponse = await authRepository.getJWTCurrent();
+    if (jwtResponse == null) {
+      log('Token is null', name: 'UserInfoRepository');
+      return ResultReturn(httpStatusCode: 401, data: null);
+    } else {
+      String jwt = jwtResponse.token;
+      Map<String, String> headers = Map.from(httpHeaders);
+      headers['Authorization'] = 'Bearer $jwt';
+
+      String authority = APIUrl.baseUrl;
+      String unencodedPath = APIUrl.pathGetAvatarUrl;
+
+      var response = await http.get(
+        Uri.http(authority, unencodedPath),
+        headers: headers,
+      );
+
+      if (response.statusCode == 200) {
+        log('Update streak successfully', name: 'UserInfoRepository');
+
+        ResponseModel responseModel = ResponseModel.fromJson(response.body);
+        String url = responseModel.data as String;
+        return ResultReturn(httpStatusCode: 200, data: url);
       } else if (response.statusCode == 401) {
         log('Token is expired', name: 'UserInfoRepository');
         await authRepository.removeJWT();
